Input some logic to grab calories from Intakes and assign it to ChartSeries when a new Intake is added, but Still cannot get chart to display.

+++ CalorieTracker/CalorieTracker/Intake+Convenience.swift
@@ -8,9 +8,10 @@
 
 import Foundation
 import CoreData
+import SwiftChart
 
 extension Intake {
-    convenience init(calories: Int32, timeStamp: Date, context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
+    convenience init(calories: Double, timeStamp: Date, context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
         self.init(context: context)
         self.calories = calories
         self.timeStamp = timeStamp
